refactor: break up giant output file of classes
