+++ src/commands/project/deploy/cancel.ts
@@ -11,7 +11,8 @@ import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
 import { RequestStatus } from '@salesforce/source-deploy-retrieve';
 import { cancelDeploy, cancelDeployAsync } from '../../../utils/deploy';
 import { DeployCache } from '../../../utils/deployCache';
-import { AsyncDeployCancelResultFormatter, DeployCancelResultFormatter } from '../../../utils/output';
+import { AsyncDeployCancelResultFormatter } from '../../../formatters/AsyncDeployCancelResultFormatter';
+import { DeployCancelResultFormatter } from '../../../formatters/DeployCancelResultFormatter';
 import { DeployResultJson } from '../../../utils/types';
 
 Messages.importMessagesDirectory(__dirname);
