+++ src/commands/project/deploy/resume.ts
@@ -9,7 +9,8 @@ import { bold } from 'chalk';
 import { EnvironmentVariable, Messages } from '@salesforce/core';
 import { SfCommand, toHelpSection, Flags } from '@salesforce/sf-plugins-core';
 import { Duration } from '@salesforce/kit';
-import { DeployResultFormatter, getVersionMessage } from '../../../utils/output';
+import { getVersionMessage } from '../../../utils/output';
+import { DeployResultFormatter } from '../../../formatters/DeployResultFormatter';
 import { DeployProgress } from '../../../utils/progressBar';
 import { DeployResultJson, reportsFormatters } from '../../../utils/types';
 import { determineExitCode, executeDeploy, isNotResumable } from '../../../utils/deploy';
