+++ src/commands/project/deploy/start.ts
@@ -8,7 +8,9 @@ import { bold } from 'chalk';
 import { EnvironmentVariable, Messages, OrgConfigProperties, SfError } from '@salesforce/core';
 import { SfCommand, toHelpSection, Flags } from '@salesforce/sf-plugins-core';
 import { SourceConflictError } from '@salesforce/source-tracking';
-import { AsyncDeployResultFormatter, DeployResultFormatter, getVersionMessage } from '../../../utils/output';
+import { getVersionMessage } from '../../../utils/output';
+import { AsyncDeployResultFormatter } from '../../../formatters/AsyncDeployResultFormatter';
+import { DeployResultFormatter } from '../../../formatters/DeployResultFormatter';
 import { DeployProgress } from '../../../utils/progressBar';
 import { DeployResultJson, TestLevel, reportsFormatters } from '../../../utils/types';
 import { executeDeploy, resolveApi, validateTests, determineExitCode } from '../../../utils/deploy';
