+++ src/commands/project/deploy/validate.ts
@@ -8,7 +8,9 @@ import { bold } from 'chalk';
 import { EnvironmentVariable, Messages, OrgConfigProperties } from '@salesforce/core';
 import { RequestStatus } from '@salesforce/source-deploy-retrieve';
 import { SfCommand, toHelpSection, Flags } from '@salesforce/sf-plugins-core';
-import { AsyncDeployResultFormatter, DeployResultFormatter, getVersionMessage } from '../../../utils/output';
+import { getVersionMessage } from '../../../utils/output';
+import { AsyncDeployResultFormatter } from '../../../formatters/AsyncDeployResultFormatter';
+import { DeployResultFormatter } from '../../../formatters/DeployResultFormatter';
 import { DeployProgress } from '../../../utils/progressBar';
 import { DeployResultJson, TestLevel } from '../../../utils/types';
 import { executeDeploy, resolveApi, determineExitCode } from '../../../utils/deploy';
