+++ src/commands/project/retrieve/start.ts
@@ -16,7 +16,8 @@ import { getString } from '@salesforce/ts-types';
 import { SourceTracking, SourceConflictError } from '@salesforce/source-tracking';
 import { Duration } from '@salesforce/kit';
 import { DEFAULT_ZIP_FILE_NAME, ensuredDirFlag, zipFileFlag } from '../../../utils/flags';
-import { MetadataRetrieveResultFormatter, RetrieveResultFormatter } from '../../../utils/output';
+import { RetrieveResultFormatter } from '../../../formatters/RetrieveResultFormatter';
+import { MetadataRetrieveResultFormatter } from '../../../formatters/MetadataRetrieveResultFormatter';
 import { getPackageDirs } from '../../../utils/project';
 import { RetrieveResultJson } from '../../../utils/types';
 import { writeConflictTable } from '../../../utils/conflicts';
